pyproject.toml
[project]
name = "nonebot-plugin-mcmotd"
version = "1.0.5"
description = "A NoneBot2 plugin for querying Minecraft server status and generating status images"
authors = [
    { name = "AquaOH", email = "aquaoh@outlook.com" }
]
dependencies = [
    "nonebot2>=2.4.0",
    "nonebot-adapter-onebot>=2.4.0",
    "nonebot-plugin-localstore>=0.4.0",
    "mcstatus>=12.0.0",
    "pillow>=10.0.0",
    "aiosqlite>=0.20.0",
    "pydantic",
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
keywords = ["nonebot", "nonebot2", "minecraft", "motd", "server-status", "qq-bot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Games/Entertainment",
    "Topic :: Internet",
]

[project.urls]
homepage = "https://github.com/AquaOH/nonebot-plugin-mcmotd"
repository = "https://github.com/AquaOH/nonebot-plugin-mcmotd"
documentation = "https://github.com/AquaOH/nonebot-plugin-mcmotd#readme"
changelog = "https://github.com/AquaOH/nonebot-plugin-mcmotd/releases"
issues = "https://github.com/AquaOH/nonebot-plugin-mcmotd/issues"

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

[build-system]
requires = ["hatchling>=1.13.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "nonebot_plugin_mcmotd/__init__.py"
pattern = '__version__ = "(?P<version>[^"]+)"'

[tool.hatch.build.targets.wheel]
packages = ["nonebot_plugin_mcmotd"]

[tool.hatch.build.targets.sdist]
include = [
    "/nonebot_plugin_mcmotd",
    "/README.md",
    "/LICENSE",
]

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 88
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "RUF012", # mutable class attributes should be annotated with typing.ClassVar
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"] # allow unused imports in __init__.py

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"